# Conflicts:
#	manim/utils/ipython_magic.py

+++ manim/utils/ipython_magic.py
@@ -6,8 +6,9 @@ import shutil
 from datetime import datetime
 from pathlib import Path
 
-from manim import config, tempconfig
+from manim import Group, config, logger, tempconfig
 from manim.__main__ import main
+from manim.renderer.shader import shader_program_cache
 
 try:
     from IPython import get_ipython
@@ -25,7 +26,7 @@ else:
     @magics_class
     class ManimMagic(Magics):
         def __init__(self, shell):
-            super(ManimMagic, self).__init__(shell)
+            super().__init__(shell)
             self.rendered_files = {}
 
         @needs_local_scope
@@ -106,13 +107,46 @@ else:
             args = main(modified_args, standalone_mode=False, prog_name="manim")
             with tempconfig(local_ns.get("config", {})):
                 config.digest_args(args)
-                exec(f"{config['scene_names'][0]}().render()", local_ns)
+
+                renderer = None
+                if config.renderer == "opengl":
+                    # Check if the imported mobjects extend the OpenGLMobject class
+                    # meaning ConvertToOpenGL did its job
+                    if "OpenGLMobject" in map(lambda cls: cls.__name__, Group.mro()):
+                        from manim.renderer.opengl_renderer import OpenGLRenderer
+
+                        renderer = OpenGLRenderer()
+                    else:
+                        logger.warning(
+                            "Renderer must be set to OpenGL in the configuration file "
+                            "before importing Manim! Using cairo renderer instead.",
+                        )
+                        config.renderer = "cairo"
+
+                try:
+                    SceneClass = local_ns[config["scene_names"][0]]
+                    scene = SceneClass(renderer=renderer)
+                    scene.render()
+                finally:
+                    # Shader cache becomes invalid as the context is destroyed
+                    shader_program_cache.clear()
+
+                    # Close OpenGL window here instead of waiting for the main thread to
+                    # finish causing the window to stay open and freeze
+                    if renderer is not None and renderer.window is not None:
+                        renderer.window.close()
+
+                if config["output_file"] is None:
+                    logger.info("No output file produced")
+                    return
+
                 local_path = Path(config["output_file"]).relative_to(Path.cwd())
                 tmpfile = (
                     Path(config["media_dir"])
                     / "jupyter"
                     / f"{_generate_file_name()}{local_path.suffix}"
                 )
+
                 if local_path in self.rendered_files:
                     self.rendered_files[local_path].unlink()
                 self.rendered_files[local_path] = tmpfile
